Add files via upload
For this project, I ran the given server code and edited the client code so that the client both allowed the user to continue to send multiple requests (instead of just one as it was originally written) stopped the program when the user typed in “Quit”. I chose to work on the python version. In the end, it was a good decision because I really only had to write 2 lines of code for each of the different client codes (for TCP version and the UDP version).